uint64 deny plumbing; P4 exact compare; OfsTest; the pin; G1 widening (guard +
ENG_ATTEMPT verbatim); bit 32 -fno-run-prefilter + run-pinned rows. Movers =
census (corpus 505 / bench 110; design's 513 was a latin-1 census bug).
make test 43/43 at 3be85839: reds = nm probe + two stale pins (registry 123->135,
resource -131 B), triaged and re-pinned (fd76693d). Sabotage S279-S289
(S284 (j), S287 (d, ASan-only) UNDETECTED-expected).


# Conflicts:
#	docs/dev/lanes/CLAUDE.md

+++ src/core/compile.c
@@ -1575,7 +1575,7 @@ static int compile_driver(const char *pattern, const pcrec_options *opt,
              * and `nst` reuses the Job-owned array, so the discarded exact
              * machine costs no allocation and — the expensive half —
              * determinization never runs on it. */
-            const unsigned pfc_flags = cx.opt->flags;
+            const uint64_t pfc_flags = cx.opt->flags;
             const bool pfc_deny  = (pfc_flags & PCREC_NO_PREFILTER_COLLAPSE) != 0;
             const bool pfc_force = (pfc_flags & PCREC_FORCE_PREFILTER_COLLAPSE) != 0;
             /* [OPT-4.1] READ, NOT RE-CALLED: `select_engine.c` derived this
